Moved DeleteIconButton and EditIconButton to common folder
